pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Steve Dignam <steve@dignam.xyz>", "Christopher Dignam <chris@dignam.xyz>"]
packages = [
    { include = "cli", from = "backend" },
    { include = "pytest_plugin", from = "backend" },
]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "^3.9.2"
django = "^2.2b1"
dj-database-url = "0.4.2"
django-allauth = "^0.39.1"
drf-nested-routers = "0.90.0"
pint = "0.8.1"
gunicorn = "19.7.1"
psycopg2 = "2.7.3.2"
python-json-logger = "0.1.8"
pyyaml = "^5.1"
django-user-sessions = "^1.6"
dataclasses = "^0.6.0"
django-softdelete = "^0.9.0"
sentry-sdk = "^0.7.2"
honcho = "^1.0"
click = "^7.0"
python-dotenv = "^0.10.1"
typing_extensions = "^3.7"

[tool.poetry.dev-dependencies]
pytest = "^3.4.0"
pytest-cov = "^2.5.1"
pytest-django = "^3.1.2"
ipdb = "^0.11.0"
mypy = "^0.700"
flake8 = "^3.6.0"
flake8-formatter-junit-xml = "^0.0.5"
black = "18.9b0"
pytest-watch = "^4.2"
mypy_extensions = "^0.4.1"
flake8-pie = "^0.2.1"
flake8-print = "^3.1"
pylint = "^2.3"
pytest-sugar = "^0.9.2"
pytest-randomly = "^2.1"
pytest-timeout = "^1.3"
pytest-testmon = "^0.9.16"
pytest-icdiff = "^0.2.0"
pywatchman = "^1.4"
isort = "^4.3"

[tool.poetry.scripts]
yak = 'cli:cli'

[tool.poetry.plugins."pytest11"]
"pytest_plugin" = "pytest_plugin.plugin"

[tool.black]
py36 = true
exclude = '''
/(
  | \.venv
  | node_modules
)/
'''


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
